fixes + rspec test for ghost hashes

+++ app/models/user_cleaner.rb
@@ -1,6 +1,6 @@
 # PORO class that removes users with inactive emails
 class UserCleaner
-  attr_accessor :email_dict, :imap
+  attr_accessor :imap, :email_dict, :hash_dict
 
   def login
     @imap = Net::IMAP.new(ENV['IMAPSERVER'], port: 993, ssl: true)
@@ -31,23 +31,39 @@ class UserCleaner
     end
   end
 
-  def hash_based_deletion
+  def send_hashes
+    @emails = @email_dict.keys
+    @users = User.where(email: @emails)
+
+    @users.each do |user|
+      user.update(ghost_hash: Digest::SHA256.hexdigest(Time.now.to_i.to_s))
+      MathiMailer.ghost_email(user).deliver_now
+      move_mail(@email_dict[user])
+    end
+  end
+
+  def find_hashes
+    @hash_dict = {}
     @imap.examine(ENV['PROJECT_EMAIL_MAILBOX'])
     @imap.search(['SUBJECT', 'Ghost']).each do |message_id|
       body = @imap.fetch(message_id, "BODY[TEXT]")[0].attr["BODY[TEXT]"]
       begin
         mail = body.match(/<p>(.*)<br>/).captures
         hash = body.match(/<br>\r\n(.*)<\/p>/).captures
+        @hash_dict[mail] = hash
       rescue
         next
       end
 
       u = User.find_by(email: mail, ghost_hash: hash)
-      if u&.generic?
-        u.destroy!
-      end
+      move_mail(message_id) if u.present?
+    end
+  end
 
-      move_mail(message_id)
+  def delete_ghosts
+    @hash_dict.each do |mail, hash|
+      u = User.find_by(email: mail, ghost_hash: hash)
+      u.destroy! if u&.generic?
     end
   end
 
@@ -61,28 +77,6 @@ class UserCleaner
       @email_dict[email] = [message_id]
     end
   end
-
-  def send_hashes
-    @emails = @email_dict.keys
-    @users = User.where(email: @emails)
-
-    @users.each do |user|
-      user.update(ghost_hash: Digest::SHA256.hexdigest(Time.now.to_i.to_s))
-      MathiMailer.ghost_email(user).deliver_now
-    end
-  end
-
-  '''def destroy_users
-    @emails = @email_dict.keys
-    @users = User.where(email: @emails)
-    @present_emails = @users.pluck(:email)
-    
-    @users.each(&:destroy!)
-
-    return if @present_emails.blank?
-    message_ids = @email_dict.values_at(*@present_emails).flatten(1).uniq
-    move_mail(message_ids)
-  end'''
 
   def move_mail(message_ids, attempt=0)
     return if message_ids.blank?
@@ -105,7 +99,8 @@ class UserCleaner
     return if @email_dict.blank?
     send_hashes
     sleep(10)
-    hash_based_deletion
+    find_hashes
+    delete_ghosts
     logout
   end
